version upgrade, bug fix, improved media error handling

+++ app/src/main/java/org/nuclearfog/twidda/backend/api/holder/MediaStream.java
@@ -39,7 +39,7 @@ public class MediaStream {
     /**
      * @return remaining bytes of the stream
      */
-    public int available() throws IOException {
+    public long available() throws IOException {
         return inputStream.available();
     }
